New TableRenderer component to render page/list table, add to cover page - Add list id to hash on scrollspy click - Don't smooth scroll to element if over 10k pixels away
