+++ frontend/src/components/collection/CollectionCoverUI/style.scss
@@ -25,7 +25,7 @@
         flex-grow: 1;
         width: calc(100% - 40px);
 
-        .overview-tab {
+        .overview-tab.react-tabs__tab-panel--selected {
             flex-grow: 1;
             display: grid;
             grid-template-rows: 100%;
@@ -58,7 +58,7 @@
                     h3 {
                         font-weight: 400;
                         margin-left: 1.5rem;
-                        font-size: 1.9;
+                        font-size: 1.9rem;
 
                         svg {
                             margin-right: 1rem;
@@ -139,7 +139,7 @@
                         }
 
                         .desc {
-                            padding: 1rem 4.5rem 0 4.5rem;
+                            padding: 1rem 0 0 2.5rem;
                             margin-bottom: 3rem;
 
                             .wr-truncate {
@@ -182,6 +182,13 @@
                     }
                 }
             }
+        }
+
+        .browse-all-tab.react-tabs__tab-panel--selected {
+            flex-grow: 1;
+            display: flex;
+            flex-direction: column;
+
         }
     }
 }
