+++ frontend/src/components/collection/CollectionDetailUI/index.js
@@ -1,52 +1,31 @@
 import React, { Component } from 'react';
 import classNames from 'classnames';
 import PropTypes from 'prop-types';
-import ArrowKeyStepper from 'react-virtualized/dist/commonjs/ArrowKeyStepper';
-import AutoSizer from 'react-virtualized/dist/commonjs/AutoSizer';
-import Column from 'react-virtualized/dist/commonjs/Table/Column';
-import Table from 'react-virtualized/dist/commonjs/Table';
 import Helmet from 'react-helmet';
 import { List } from 'immutable';
-import { Button } from 'react-bootstrap';
 
 import config from 'config';
 
 import { setSort } from 'store/modules/collection';
-import { getCollectionLink, getListLink, getStorage, inStorage, range, setStorage, truncate } from 'helpers/utils';
+import { getCollectionLink, getListLink, range, truncate } from 'helpers/utils';
 
 import {
-  CollectionFilters,
   CollectionHeader,
   InspectorPanel,
   Lists,
-  ListHeader,
   Sidebar,
   Temp404,
   TempUserAlert
 } from 'containers';
 
 import HttpStatus from 'components/HttpStatus';
-import Modal from 'components/Modal';
-import OutsideClick from 'components/OutsideClick';
 import RedirectWithStatus from 'components/RedirectWithStatus';
 import Resizable from 'components/Resizable';
+import TableRenderer from 'components/collection/TableRenderer';
 
 import 'react-virtualized/styles.css';
 
 import CustomDragLayer from './dragLayer';
-import { DefaultRow, DnDRow, DnDSortableRow } from './rows';
-import {
-  BrowserRenderer,
-  DefaultHeader,
-  DnDSortableHeader,
-  LinkRenderer,
-  RemoveRenderer,
-  RowIndexRenderer,
-  SessionRenderer,
-  TitleRenderer,
-  TimestampRenderer
-} from './columns';
-
 import './style.scss';
 
 
@@ -81,7 +60,7 @@ class CollectionDetailUI extends Component {
     setPageInspector: PropTypes.func
   };
 
-  constructor(props, context) {
+  constructor(props) {
     super(props);
 
     this.keyBuffer = List();
@@ -98,23 +77,12 @@ class CollectionDetailUI extends Component {
       },
     ];
 
-    this.columns = config.columns;
-    this.initialState = {
-      columns: context.isMobile ? ['timestamp', 'url'] : config.defaultColumns,
-      headerEditor: false,
+    this.state = {
       listBookmarks: props.list.get('bookmarks'),
       sortedBookmarks: props.list.get('bookmarks'),
       overrideHeight: null,
       selectedPageIdx: null
     };
-
-    const activeList = Boolean(props.match.params.list);
-
-    if (activeList && context.canAdmin && !context.isMobile) {
-      this.initialState.columns = ['remove', ...this.initialState.columns];
-    }
-
-    this.state = this.initialState;
   }
 
   componentWillMount() {
@@ -122,29 +90,9 @@ class CollectionDetailUI extends Component {
   }
 
   componentDidMount() {
-    const { canAdmin, isMobile } = this.context;
-
-    if (isMobile) {
-      return false;
+    if (!this.context.isMobile) {
+      document.addEventListener('keydown', this.handleKeyInput);
     }
-
-    if (inStorage('columnOrder')) {
-      try {
-        const columns = JSON.parse(getStorage('columnOrder')).filter(o => config.columns.includes(o));
-
-        if ((!this.props.match.params.list || !canAdmin) && columns.includes('remove')) {
-          columns.splice(columns.indexOf('remove'), 1);
-        } else if (this.props.match.params.list && canAdmin && !columns.includes('remove')) {
-          columns.unshift('remove');
-        }
-
-        this.setState({ columns });
-      } catch (e) {
-        console.log('Wrong `columnOrder` storage value.', e);
-      }
-    }
-
-    document.addEventListener('keydown', this.handleKeyInput);
   }
 
   componentWillReceiveProps(nextProps) {
@@ -178,7 +126,9 @@ class CollectionDetailUI extends Component {
   }
 
   componentWillUnmount() {
-    document.removeEventListener('keydown', this.handleKeyInput);
+    if (!this.context.isMobile) {
+      document.removeEventListener('keydown', this.handleKeyInput);
+    }
   }
 
   onKeyNavigate = ({ scrollToRow }) => {
@@ -304,16 +254,6 @@ class CollectionDetailUI extends Component {
     }
   }
 
-  testRowHighlight = ({ index }) => {
-    const { selectedPageIdx } = this.state;
-    const baseClass = index % 2 !== 0 ? 'odd' : '';
-
-    if (!!selectedPageIdx && typeof selectedPageIdx === 'object') {
-      return selectedPageIdx.includes(index) ? `${baseClass} selected` : baseClass;
-    }
-    return index === selectedPageIdx ? `${baseClass} selected` : baseClass;
-  }
-
   sortBookmark = (origIndex, hoverIndex) => {
     const { sortedBookmarks } = this.state;
     const o = sortedBookmarks.get(origIndex);
@@ -329,70 +269,6 @@ class CollectionDetailUI extends Component {
     const order = sortedBookmarks.map(o => o.get('id')).toArray();
     this.setState({ listBookmarks: sortedBookmarks });
     saveBookmarkSort(list.get('id'), order);
-  }
-
-  orderColumn = (origIndex, hoverIndex) => {
-    const { columns } = this.state;
-    const c = columns.splice(origIndex, 1)[0];
-    columns.splice(hoverIndex, 0, c);
-    this.setState({ columns });
-    this.saveHeaderState();
-  }
-
-  saveHeaderState = () => {
-    setStorage('columnOrder', JSON.stringify(this.state.columns));
-  }
-
-  toggleHeaderModal = () => {
-    this.setState({ headerEditor: !this.state.headerEditor });
-  }
-
-  toggleColumn = (evt) => {
-    const { columns } = this.state;
-    const idx = columns.indexOf(evt.target.name);
-
-    if (idx !== -1) {
-      columns.splice(idx, 1);
-    } else {
-      columns.push(evt.target.name);
-    }
-
-    this.setState({ columns });
-    this.saveHeaderState();
-  }
-
-  customRowRenderer = (props) => {
-    const { canAdmin } = this.context;
-    const { match: { params: { list } } } = this.props;
-
-    if (!canAdmin || __PLAYER__) {
-      return <DefaultRow {...props} />;
-    }
-
-    if (list) {
-      return (
-        <DnDSortableRow
-          id={props.rowData.get('id')}
-          save={this.saveSort}
-          sort={this.sortBookmark}
-          pageSelection={this.state.selectedPageIdx}
-          {...props} />
-      );
-    }
-    return <DnDRow {...props} />;
-  }
-
-  customHeaderRenderer = (props) => {
-    if (__PLAYER__) {
-      return <DefaultHeader {...props} />;
-    }
-
-    return (
-      <DnDSortableHeader
-        save={this.saveCollOrder}
-        order={this.orderColumn}
-        {...props} />
-    );
   }
 
   collapsibleToggle = (isOpen) => {
@@ -441,77 +317,6 @@ class CollectionDetailUI extends Component {
     const indexPages = !canAdmin && !publicIndex ? List() : pages;
     const objects = activeList ? listBookmarks : indexPages;
     const displayObjects = activeList ? sortedBookmarks : indexPages;
-    const objectLabel = activeList ? 'Bookmark Title' : 'Page Title';
-
-    const columnDefs = {
-      browser: {
-        cellRenderer: BrowserRenderer,
-        columnData: { browsers },
-        dataKey: 'browser',
-        key: 'browser',
-        width: 150
-      },
-      rowIndex: {
-        cellRenderer: RowIndexRenderer,
-        dataKey: 'id',
-        disableSort: true,
-        key: 'rowIndex',
-        width: 60
-      },
-      remove: {
-        cellRenderer: RemoveRenderer,
-        columnData: {
-          bkDeleting: this.props.bkDeleting,
-          bkDeleteError: this.props.bkDeleteError,
-          listId: activeList ? list.get('id') : null,
-          removeCallback: this.props.removeBookmark
-        },
-        dataKey: 'remove',
-        key: 'remove',
-        width: 55
-      },
-      session: {
-        cellRenderer: SessionRenderer,
-        columnData: {
-          activeList,
-          canAdmin,
-          collLink: getCollectionLink(collection)
-        },
-        dataKey: 'rec',
-        key: 'session',
-        width: 100
-      },
-      timestamp: {
-        cellRenderer: TimestampRenderer,
-        dataKey: 'timestamp',
-        key: 'timestamp',
-        width: 200
-      },
-      title: {
-        cellRenderer: TitleRenderer,
-        className: 'page-title',
-        columnData: {
-          collection,
-          list: activeList ? list : null
-        },
-        dataKey: 'title',
-        flexGrow: 1,
-        key: 'title',
-        label: objectLabel,
-        width: 200
-      },
-      url: {
-        cellRenderer: LinkRenderer,
-        columnData: {
-          collection,
-          list: activeList ? list : null
-        },
-        dataKey: 'url',
-        flexGrow: 1,
-        key: 'url',
-        width: 200
-      }
-    };
 
     return (
       <div className={classNames('wr-coll-detail', { 'with-list': activeList })}>
@@ -573,108 +378,20 @@ class CollectionDetailUI extends Component {
           </div>
         </Sidebar>
 
-
-        <div className="table-container">
-          {
-            activeList ?
-              <ListHeader /> :
-              <div className="collection-header">
-                <h2>Pages</h2>
-                <CollectionFilters />
-              </div>
-          }
-
-          <OutsideClick classes="wr-coll-detail-table" handleClick={this.deselect}>
-            {
-              canAdmin &&
-                <React.Fragment>
-                  <Button onClick={this.toggleHeaderModal} className="table-header-menu borderless" bsSize="xs">
-                    {/* TODO: placeholder icon */}
-                    <span style={{ display: 'inline-block', fontWeight: 'bold', transform: 'rotateZ(90deg)' }}>...</span>
-                  </Button>
-                  <Modal
-                    visible={this.state.headerEditor}
-                    closeCb={this.toggleHeaderModal}
-                    dialogClassName="table-header-modal"
-                    header={<h4>Edit Table Columns</h4>}
-                    footer={<Button onClick={this.toggleHeaderModal}>Close</Button>}>
-                    <ul>
-                      {
-                        this.columns.map((coll) => {
-                          return (
-                            <li key={coll}>
-                              <input type="checkbox" onChange={this.toggleColumn} name={coll} id={`add-to-list-${coll}`} checked={this.state.columns.includes(coll) || false} />
-                              <label htmlFor={`add-to-list-${coll}`}>{config.columnLabels[coll] || coll}</label>
-                            </li>
-                          );
-                        })
-                      }
-                    </ul>
-                  </Modal>
-                </React.Fragment>
-            }
-            <AutoSizer>
-              {
-                ({ height, width }) => (
-                  <ArrowKeyStepper
-                    rowCount={displayObjects.size}
-                    columnCount={1}
-                    mode="cells"
-                    scrollToRow={typeof selectedPageIdx === 'number' ? selectedPageIdx : 0}
-                    onScrollToChange={this.onKeyNavigate}>
-                    {
-                      ({ onSectionRendered, scrollToRow }) => {
-                        return (
-                          <Table
-                            width={width}
-                            height={height}
-                            rowCount={displayObjects.size}
-                            headerHeight={25}
-                            rowHeight={40}
-                            rowGetter={({ index }) => displayObjects.get(index)}
-                            rowClassName={this.testRowHighlight}
-                            onRowClick={this.onSelectRow}
-                            onRowsRendered={({ startIndex, stopIndex }) => {
-                              onSectionRendered({ rowStartIndex: startIndex, rowStopIndex: stopIndex })
-                            }}
-                            rowRenderer={this.customRowRenderer}
-                            sort={activeList ? null : this.sort}
-                            sortBy={activeList ? '' : collection.getIn(['sortBy', 'sort'])}
-                            sortDirection={activeList ? null : collection.getIn(['sortBy', 'dir'])}>
-                            {
-                              this.state.columns.map((c, idx) => {
-                                let props = columnDefs[c];
-                                let collData = {};
-
-                                if (props.hasOwnProperty('columnData')) {
-                                  props = Object.assign({}, props);
-                                  collData = props.columnData;
-                                  delete props.columnData;
-                                }
-
-                                if (!props.label) {
-                                  props.label = config.columnLabels[[c]] || c;
-                                }
-
-                                return (
-                                  <Column
-                                    headerRenderer={this.customHeaderRenderer}
-                                    index={idx}
-                                    columnData={{ ...collData, index: idx }}
-                                    {...props} />
-                                );
-                              })
-                            }
-                          </Table>
-                        );
-                      }
-                    }
-                  </ArrowKeyStepper>
-                )
-              }
-            </AutoSizer>
-          </OutsideClick>
-        </div>
+        <TableRenderer {...{
+          activeList,
+          browsers,
+          collection,
+          deselect: this.deselect,
+          displayObjects,
+          list,
+          onKeyNavigate: this.onKeyNavigate,
+          onSelectRow: this.onSelectRow,
+          saveSort: this.saveSort,
+          selectedPageIdx,
+          sort: this.sort,
+          sortBookmark: this.sortBookmark
+        }} />
       </div>
     );
   }
