+++ frontend/src/components/collection/TableRenderer/style.scss
@@ -1,0 +1,197 @@
+@import 'src/vars';
+
+.table-container {
+    display: flex;
+    flex-direction: column;
+    flex-grow: 1;
+
+    .collection-header {
+        display: flex;
+        justify-content: space-between;
+        align-items: center;
+
+        h2 {
+            font-size: 2.8rem;
+            font-weight: 400;
+            color: $gray600;
+            margin-left: 1.5rem;
+        }
+
+
+        .wr-coll-utilities {
+            background: #FFF;
+            //z-index: 10;
+            margin-left: 1.5rem;
+            width: 30%;
+            min-width: 25rem;
+            max-width: 35rem;
+
+            nav {
+                height: 100%;
+                display: flex;
+                justify-content: flex-start;
+                align-items: center;
+
+                .search-box {
+                    width: 100%;
+                    margin-right: .5rem;
+                }
+            }
+        }
+    }
+
+    .wr-coll-detail-table {
+        flex-grow: 1;
+        position: relative;
+        padding-right: 1.5rem;
+        padding-left: 1.5rem;
+        background: #FFF;
+
+        .table-header-menu {
+            position: absolute;
+            right: -.5rem;
+            top: 0.25rem;
+        }
+
+        .ReactVirtualized__Table__headerRow {
+            font-size: 1rem;
+            font-weight: 500;
+            letter-spacing: 0.5pt;
+        }
+
+        .ReactVirtualized__Table__headerColumn {
+            height: 100%;
+
+            &:first-of-type {
+                margin-left: 0;
+            }
+
+            &:last-of-type {
+                margin-right: 0;
+            }
+
+            & > div{
+                display: flex;
+                align-items: center;
+                background: $tableHeaderBg;
+                padding: 0 .5rem 0 0;
+                width: 100%;
+                height: 100%;
+            }
+
+            .header-handle {
+                width: 1rem;
+                height: 100%;
+                margin-right: .75rem;
+                cursor: move;
+                background-image: url('../../../shared/images/Table-Col-Header-Handle.svg');
+                background-repeat: no-repeat;
+                flex-shrink: 0;
+                background-position: .3rem .6rem;
+            }
+        }
+
+        .ReactVirtualized__Table__row {
+            transition: background $tableRowFade ease-in-out,
+                        color $tableRowFade ease-in-out,
+                        fill $tableRowFade ease-in-out;
+            cursor: context-menu;
+            font-size: 1.2rem;
+            position: relative;
+            background: #FFF;
+
+            &:before {
+                transition: background-color $quickShift ease-in-out;
+                position: absolute;
+                left: 0;
+                content: '';
+                width: .75rem;
+                height: 100%;
+            }
+
+            &:focus {
+                outline: none;
+            }
+
+            &.odd {
+                background-color: $table-zebra;
+            }
+
+            .ReactVirtualized__Table__rowColumn {
+                margin-left: 0;
+
+                &:first-of-type {
+                    & > * {
+                        padding: 0 1.25rem;
+                    }
+                }
+
+                .wr-remove-widget {
+                    text-align: left;
+
+                    button {
+                        margin: 0;
+                    }
+                }
+
+                .session-link {
+                    &:link, &:visited {
+                        transition: color $tableRowFade ease-in-out;
+                        color: #333;
+                    }
+                }
+
+                .row-index {
+                    text-align: center;
+                }
+            }
+
+            .page-title {
+                font-size: 1.45rem;
+            }
+
+            svg {
+                & > g {
+                    transition: fill $tableRowFade ease-in-out;
+                }
+            }
+
+            &:hover {
+                &:not(.selected){
+                    &:before {
+                        background: $tr-selected;
+                    }
+                }
+
+                &.selected {
+                    &:before {
+                        background: $tr-highlight-selected;
+                    }
+                }
+            }
+
+            &.selected {
+                background-color: $blue;
+                color: #FFF;
+                fill: #FFF;
+
+                .session-link:link, a:link { color: #FFF; }
+                .session-link:visited, a:visited{ color: #FFF; }
+
+                svg {
+                    & > g {
+                        fill: #FFF;
+                    }
+                }
+            }
+
+
+            .ReactVirtualized__Table__rowColumn {
+                a {
+                    display: inline-block;
+                    padding: 1rem;
+                }
+            }
+        }
+    }
+}
