+++ frontend/src/containers/CollectionCover/CollectionCover.js
@@ -5,6 +5,9 @@ import { asyncConnect } from 'redux-connect';
 import { timestampOrderedPages } from 'store/selectors';
 
 import { isLoaded as isCollLoaded, load as loadColl, loadLists } from 'store/modules/collection';
+import { isLoaded as isRBLoaded, load as loadRB } from 'store/modules/remoteBrowsers';
+import { getQueryPages, getOrderedPages } from 'store/selectors';
+import { pageSearchResults } from 'store/selectors/search';
 
 import CollectionCoverUI from 'components/collection/CollectionCoverUI';
 
@@ -53,14 +56,41 @@ const initialData = [
         .then(() => dispatch(loadLists(user, coll, 'all')));
     }
   },
+  {
+    promise: ({ store: { dispatch, getState } }) => {
+      const state = getState();
+
+      if (!isRBLoaded(state) && !__PLAYER__) {
+        return dispatch(loadRB());
+      }
+
+      return undefined;
+    }
+  }
 ];
 
 
-const mapStateToProps = ({ app }) => {
+const mapStateToProps = (outerState) => {
+  const { app } = outerState;
+  const isLoaded = app.getIn(['collection', 'loaded']);
+  const { pageFeed, searchText } = isLoaded ? pageSearchResults(outerState) : { pageFeed: Map(), searchText: '' };
+  const isIndexing = isLoaded && !pageFeed.size && app.getIn(['collection', 'pages']).size && !searchText;
+
+  const querying = app.getIn(['pageQuery', 'querying']);
+  let pages;
+
+  if (querying) {
+    pages = getQueryPages(app);
+  } else {
+    pages = isIndexing ? getOrderedPages(app) : pageFeed;
+  }
+
   return {
     auth: app.get('auth'),
+    browsers: app.get('remoteBrowsers'),
     collection: app.get('collection'),
-    orderdPages: timestampOrderedPages(app)
+    orderdPages: timestampOrderedPages(app),
+    pages
   };
 };
